5차 정리

+++ calorie/views.py
@@ -1,44 +1,69 @@
 import io
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.db import transaction
-from django.http import HttpResponseRedirect
+from django.db.models.functions import Round
 from django.shortcuts import render, redirect
-from django.utils import timezone
 from django.contrib.auth.decorators import login_required
 import random
 from django.core.paginator import Paginator
 from . import models
 import requests
 import base64
-import time
 from django.core.files import File
+from datetime import datetime
+from django.db.models import Sum
+
+from .models import UserFoodTime
 
 
 @login_required
 def mypage(request, date=None):
     user = request.user
-    date = request.GET.get('date') or timezone.now().strftime("%Y-%m-%d")
+    now = datetime.now(tz=None)
+    date = request.GET.get('date') or now.strftime("%Y-%m-%d")
 
     if request.method == 'POST':
-        memo_text = request.POST.get('memo_text', '')
-        print(memo_text)
+        if 'memo_text' in request.POST:
+            memo_text = request.POST.get('memo_text', '')
 
-        if memo_text:
-            memos, created = models.UserMemo.objects.get_or_create(user_id=user.id, created_at=date,
-                                                                   defaults={'description': memo_text})
-            if not created:
-                memos.description = memo_text
-                memos.save()
-        else:
-            memos = models.UserMemo.objects.filter(user_id=user.id, created_at=date).first()
-            if memos:
-                memos.delete()
-        redirect('calorie:mypage_with_date', date=date)
+            if memo_text:
+                memos, created = models.UserMemo.objects.get_or_create(user_id=user.id, created_at=date,
+                                                                       defaults={'description': memo_text})
+                if not created:
+                    memos.description = memo_text
+                    memos.save()
+            else:
+                memos = models.UserMemo.objects.filter(user_id=user.id, created_at=date).first()
+                if memos:
+                    memos.delete()
+            redirect('calorie:mypage_with_date', date=date)
+
+    food_date = datetime.strptime(date, '%Y-%m-%d')
+    food_date = food_date.strftime('%Y-%m-%d')
 
     memos = models.UserMemo.objects.filter(user_id=user.id, created_at=date).first()
-    foods = models.UserFood.objects.filter(user_id=user.id, created_at=date).first()
+    foods = models.UserFood.objects.filter(user_id=user.id,
+                                           created_at__year=food_date[:4],
+                                           created_at__month=food_date[5:7],
+                                           created_at__day=food_date[8:10]
+                                           ).prefetch_related('user_food_images', 'user_food_time').all()
 
-    return render(request, "mypage.html", {"user": user, "memos": memos, "date": date, "foods": foods})
+    food_nutri = foods.aggregate(
+        total_energy=Round(Sum('energy'), 1),
+        total_carbohydrate=Round(Sum('carbohydrate'), 1),
+        total_protein=Round(Sum('protein'), 1),
+        total_fat=Round(Sum('fat'), 1)
+    )
+
+    if not foods:
+        food_time = 'N/A'
+    else:
+        try:
+            food_time = foods[0].user_food_time.get().food_time
+        except UserFoodTime.DoesNotExist:
+            food_time = ''
+    return render(request, "mypage.html", {"user": user, "memos": memos, "date": date,
+                                           "foods": foods, "food_nutri": food_nutri, 'food_time': food_time})
 
 
 @login_required
@@ -59,17 +84,8 @@ def service(request):
 
             if response.status_code == 200:
                 # FastAPI에서 반환한 JSON에서 이미지와 결과를 추출하고 저장
-                image_with_results = None
-                while True:
-                    try:
-                        image_with_results = response.json()
-                        print("모델 분석 완료")
-                        break
-                    except:
-                        # 결과가 아직 도착하지 않았으면 1초 대기 후 다시 요청
-                        time.sleep(1)
-                        response = requests.post(url, files=uploaded_files)
-                        continue
+                image_with_results = response.json()
+                # print("모델 분석 완료")
 
                 if image_with_results is not None:
                     user = request.user
@@ -88,23 +104,24 @@ def service(request):
                         # BytesIO 객체화
                         image_io = io.BytesIO(decoded_img)
                         # InMemoryUploadedFile 객체 생성(메모리에서 생성된 파일 객체)
+                        now = datetime.now(tz=None)
                         image_file = InMemoryUploadedFile(image_io,
                                                           None,
                                                           f'{user_name}-'
                                                           f'{food.name}-'
-                                                          f'{timezone.now().strftime("%Y-%m-%d")}.jpeg',
+                                                          f'{now.strftime("%Y-%m-%d")}.jpeg',
                                                           'image/jpeg',
                                                           image_io.getbuffer().nbytes,
                                                           None
                                                           )
                         # InMemoryUploadedFile 객체를 File 객체로 변환
                         image_file = File(image_file)
-                        print(image_file)
 
                         with transaction.atomic():
                             user_food_obj = models.UserFood.objects.create(user_id=user.id,
                                                                            name_id=food.id,
                                                                            weight=food.weight,
+                                                                           energy=food.energy,
                                                                            carbohydrate=food.carbohydrate,
                                                                            protein=food.protein,
                                                                            fat=food.fat)
@@ -115,16 +132,39 @@ def service(request):
                         user_food_list.append(confirm_food)
                     return render(request, 'service.html', {"user_food_list": user_food_list})
 
-        elif 'user_weight' in request.POST:
-            print("여기로 왔나?")
-            #     user = request.user
-            #     weight = request.POST["user_weight"]
-            #     if weight:
-            #         models.UserFood.objects.create(weight=weight, user=user)
-            #
-            # user_foods = models.UserFood.objects.filter(user_id=user.id).all()
-            # user_food_img = models.UserFoodImage.objects.filter(name=user_foods).all()
-            # return render(request, "service.html", {"user_foods": user_foods, "user_food_img": user_food_img})
+        elif 'user_weight_1' in request.POST:
+            # user = request.user
+            for i in range(1, (len(request.POST) - 1) // 3 + 1):
+                weight = request.POST.get(f"user_weight_{i}")
+                created_at_str = request.POST.get(f"created_at_{i}")
+                food_name = request.POST.get(f"food_name_{i}")
+                # AM/PM 표시를 영어로 변환
+                created_at_str = created_at_str.replace("오전", "AM").replace("오후", "PM")
+                # # 문자열을 datetime 객체로 변환
+                created_at_dt = datetime.strptime(created_at_str, '%Y년 %m월 %d일 %I:%M %p')
+                # # datetime 객체를 원하는 형식의 문자열로 변환
+                created_at_formatted = created_at_dt.strftime('%Y-%m-%d %H:%M:%S')
+
+                food = models.Food.objects.filter(name=food_name).first()
+                user_food = models.UserFood.objects.filter(
+                    name_id=food.id,
+                    created_at__year=created_at_formatted[:4],
+                    created_at__month=created_at_formatted[5:7],
+                    created_at__day=created_at_formatted[8:10],
+                    created_at__hour=created_at_formatted[11:13],
+                    created_at__minute=created_at_formatted[14:16],
+                ).first()
+                if user_food:
+                    user_food.weight = weight
+                    user_food.energy = f'{food.energy * (float(weight) / 100):.1f}'
+                    user_food.carbohydrate = f'{food.carbohydrate * (float(weight) / 100):.1f}'
+                    user_food.protein = f'{food.protein * (float(weight) / 100):.1f}'
+                    user_food.fat = f'{food.fat * (float(weight) / 100):.1f}'
+                    user_food.save()
+                    food_time = request.POST.get('food_time')
+                    models.UserFoodTime.objects.create(food_time=food_time, name_id=user_food.id)
+
+            return redirect('calorie:mypage')
 
     return render(request, "service.html")
